+++ Potentiometer.h
@@ -4,5 +4,5 @@
 #define ADC_res      12     // Use 12 bits for maximum resolution
 #define ADC_maxvolts 3.3
 
-uint16_t potRead(void);
-uint16_t potLimit(uint16_t);
+float potRead(void);
+float potLimit(float);
